place order back end

+++ shoe-shop-back-end/src/main/java/lk/ijse/gdse/shoeshopbackend/dto/OrderDTO.java
@@ -14,7 +14,7 @@ public class OrderDTO {
     private String orderId;
     private Timestamp orderDate;
     private Double totalPrice;
-    private Double addedPoints;
+    private Integer addedPoints;
     private String paymentMethod;
     private String cashierName;
     private String customer_id;
